pyproject.toml
[project]
name = "vertebrae"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "build>=1.2.2.post1",
    "collection>=0.1.6",
    "ipywidgets>=8.1.5",
    "jupyterlab-widgets>=3.0.13",
    "matplotlib>=3.9.3",
    "nibabel>=5.3.2",
    "numpy>=2.2.0",
    "opencv-python>=4.11.0.86",
    "optuna[visualization]>=4.3.0",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "scikit-image>=0.25.0",
    "scikit-learn>=1.6.0",
    "scikit-optimize>=0.10.2",
    "scipy>=1.14.1",
    "setuptools>=75.7.0",
    "simpleitk>=2.4.1",
    "torch==2.4.0+cu121",
    "torchvision==0.19.0+cu121",
    "tqdm>=4.67.1",
    "umap-learn>=0.5.7",
    "wandb>=0.15.12",
    "wheel>=0.45.1",
    "connected-components-3d>=3.12.0",
    "cupy-cuda12x>=13.0.0",
    "numba>=0.59.0",
    "ultralytics>=8.0.0",
    "albumentations>=1.4.0",
    "seaborn>=0.12.0",
    "japanize-matplotlib>=1.1.3",
    "hydra-core>=1.3.2",
    "pytorch-lightning>=2.5.5",
]

[tool.uv]
find-links = [
    "https://download.pytorch.org/whl/cu121/torch",
    "https://download.pytorch.org/whl/cu121/torchvision",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
